requirements.txt
streamlit==1.40.1
pandas==2.1.4
numpy==1.26.4
streamlit-custom-notification-box==0.1.2
streamlit-extras==0.5.5
python-dotenv==1.0.0
